Edits
Arranged the code into files for making it cleaner
